+++ pkg/cep_ent/gpu/where.go
@@ -102,7 +102,7 @@ func Memory(v int) predicate.Gpu {
 }
 
 // LowestEarnMonth applies equality check predicate on the "lowest_earn_month" field. It's identical to LowestEarnMonthEQ.
-func LowestEarnMonth(v int) predicate.Gpu {
+func LowestEarnMonth(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldEQ(FieldLowestEarnMonth, v))
 }
 
@@ -497,42 +497,42 @@ func MemoryLTE(v int) predicate.Gpu {
 }
 
 // LowestEarnMonthEQ applies the EQ predicate on the "lowest_earn_month" field.
-func LowestEarnMonthEQ(v int) predicate.Gpu {
+func LowestEarnMonthEQ(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldEQ(FieldLowestEarnMonth, v))
 }
 
 // LowestEarnMonthNEQ applies the NEQ predicate on the "lowest_earn_month" field.
-func LowestEarnMonthNEQ(v int) predicate.Gpu {
+func LowestEarnMonthNEQ(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldNEQ(FieldLowestEarnMonth, v))
 }
 
 // LowestEarnMonthIn applies the In predicate on the "lowest_earn_month" field.
-func LowestEarnMonthIn(vs ...int) predicate.Gpu {
+func LowestEarnMonthIn(vs ...int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldIn(FieldLowestEarnMonth, vs...))
 }
 
 // LowestEarnMonthNotIn applies the NotIn predicate on the "lowest_earn_month" field.
-func LowestEarnMonthNotIn(vs ...int) predicate.Gpu {
+func LowestEarnMonthNotIn(vs ...int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldNotIn(FieldLowestEarnMonth, vs...))
 }
 
 // LowestEarnMonthGT applies the GT predicate on the "lowest_earn_month" field.
-func LowestEarnMonthGT(v int) predicate.Gpu {
+func LowestEarnMonthGT(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldGT(FieldLowestEarnMonth, v))
 }
 
 // LowestEarnMonthGTE applies the GTE predicate on the "lowest_earn_month" field.
-func LowestEarnMonthGTE(v int) predicate.Gpu {
+func LowestEarnMonthGTE(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldGTE(FieldLowestEarnMonth, v))
 }
 
 // LowestEarnMonthLT applies the LT predicate on the "lowest_earn_month" field.
-func LowestEarnMonthLT(v int) predicate.Gpu {
+func LowestEarnMonthLT(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldLT(FieldLowestEarnMonth, v))
 }
 
 // LowestEarnMonthLTE applies the LTE predicate on the "lowest_earn_month" field.
-func LowestEarnMonthLTE(v int) predicate.Gpu {
+func LowestEarnMonthLTE(v int64) predicate.Gpu {
 	return predicate.Gpu(sql.FieldLTE(FieldLowestEarnMonth, v))
 }
